added file protection feature

+++ clear.py
@@ -3,6 +3,12 @@ import shutil
 
 import_dir = "./.import"
 extensions = [".tscn-", ".tmp", ".cfg", ".save", ".dmp"]
+protected_files = [
+    "global_script_class_cache.cfg", 
+    "project_metadata.cfg", 
+    "script_editor_cache.cfg", 
+    "editor_layout.cfg"
+    ]
 
 if os.path.exists(import_dir):
     shutil.rmtree(import_dir)
@@ -10,6 +16,10 @@ if os.path.exists(import_dir):
 
 for root, dirs, files in os.walk("."):
     for file in files:
+        if file in protected_files:
+            print(f"Protected file, not removed: {file}")
+            continue
+
         if any(file.endswith(ext) for ext in extensions):
             file_path = os.path.join(root, file)
             os.remove(file_path)
